pyproject.toml
[project]
name = "cover-letter-tweaker"
version = "0.1.0"
description = "FastAPI app to help tweak cover letters for different job postings"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.2",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.20",
    "python-docx>=1.2.0",
    "black>=25.11.0",
    "google-api-core>=2.28.1",
    "python-magic>=0.4.27",
    "trafilatura>=1.12.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "N",   # pep8-naming
]
ignore = [
    "N802", # Allow camelCase function names (for FastAPI route parameters)
    "N803", # Allow camelCase arguments (for FastAPI route parameters)
    "E501", # Line too long (we'll rely on formatter for this)
]

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod"]
